Login y registro funcionales

+++ database/migrations/2020_05_24_190000_create_users_table.php
@@ -24,7 +24,7 @@ class CreateUsersTable extends Migration
             $table->string('address');
             $table->string('city');
             $table->string('province');
-            $table->integer('cc_number')->nullable();
+            $table->bigInteger('cc_number')->nullable();
             $table->bigInteger('role')->unsigned();
             $table->rememberToken();
             $table->timestamps();
